calabash can click links in scaled webview
tldr; HTML pages being scaled for your device should specify a viewport
for you mobile device. Something like:

<meta name="viewport" content="width=device-width; initial-scale=1.0;
maximum-scale=1.0; user-scalable=0;" />

I was able to get around this by adding a static footer to the
scaffolded HTML.  I noticed that the link itself resizes when there is
additional content on the page.

That is, when the link text is the only thing on the page, it is
rendered smaller than when there are other things on the page. However,
the coordinates of the link obtained from querying the view hierarchy do
not correspond to the coordinates of the view in the actual interface,
but rather to the size it will be upon resizing.